@@ -20,9 +20,9 @@
         "illuminate/view": "5.5.*|5.6.*|5.7.*|5.8.x"
     },
     "require-dev": {
-        "laravel/framework": "5.0.*",
+        "laravel/framework": "5.5.*",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~4.0",
+        "phpunit/phpunit": "~6.0",
         "squizlabs/php_codesniffer": "~1.5",
         "satooshi/php-coveralls": "~0.6"
     },
