@@ -22,7 +22,8 @@ use Twig\Extension\AbstractExtension;
  * Dump a variable or the view context
  *
  * Based on the Symfony Twig Bridge Dump Extension
- * @see https://github.com/symfony/symfony/blob/2.6/src/Symfony/Bridge/Twig/Extension/DumpExtension.php
+ *
+ * @see    https://github.com/symfony/symfony/blob/2.6/src/Symfony/Bridge/Twig/Extension/DumpExtension.php
  * @author Nicolas Grekas <p@tchwork.com>
  */
 class Dump extends AbstractExtension
@@ -34,9 +35,13 @@ class Dump extends AbstractExtension
 
     public function getFunctions()
     {
-        return array(
-          new TwigFunction('dump', array($this, 'dump'), array('is_safe' => array('html'), 'needs_context' => true, 'needs_environment' => true)),
-        );
+        return [
+            new TwigFunction('dump', [$this, 'dump'], [
+                'is_safe'           => ['html'],
+                'needs_context'     => true,
+                'needs_environment' => true
+            ]),
+        ];
     }
 
     public function getName()
@@ -46,17 +51,17 @@ class Dump extends AbstractExtension
 
     public function dump(Environment $env, $context)
     {
-        if (!$env->isDebug()) {
+        if (! $env->isDebug()) {
             return;
         }
         if (2 === func_num_args()) {
-            $vars = array();
+            $vars = [];
             foreach ($context as $key => $value) {
-                if (!$value instanceof Template) {
+                if (! $value instanceof Template) {
                     $vars[$key] = $value;
                 }
             }
-            $vars = array($vars);
+            $vars = [$vars];
         } else {
             $vars = func_get_args();
             unset($vars[0], $vars[1]);
@@ -67,6 +72,7 @@ class Dump extends AbstractExtension
             $dumper->dump($this->cloner->cloneVar($value));
         }
         rewind($dump);
+
         return stream_get_contents($dump);
     }
 }
